toutes-les-entites-routes-ok

+++ web/services/ViewEngine.php
@@ -6,49 +6,61 @@ class ViewEngine {
 
         public function generateTable($data, $path){
             
-            $table = '<div class="table-responsive">
-                        <table class="table table-bordered">
-                        <thead>
-                            <tr>';
-            foreach($data[0] as $key => $value){
-                $table = $table . '<th scope="col">'. $key . '</th>';
-            }
+            $table = "";
 
-            $table = $table . '<th scope="col">Actions</th>';
-            $table = $table . '</tr></thead>';
+            if (!isset($data) || !$data || count($data) == 0){
+                $table = '<div class="alert alert-warning" role="alert">
+                            Pas de données-tableau vide
+                        </div>';
+            } else {
+                $table = '<div class="table-responsive">
+                <table class="table table-bordered">
+                <thead>
+                    <tr>';
 
-            $table = $table . '<tbody>';
-            
-            foreach($data as $key => $obs){
+                foreach($data[0] as $key => $value){
+                    $table = $table . '<th scope="col">'. $key . '</th>';
+                }
 
-                $table = $table . '<tr>';
+                $table = $table . '<th scope="col">Actions</th>';
+                $table = $table . '</tr></thead>';
+
+                $table = $table . '<tbody>';
                 
-                foreach($obs as $key => $value){
-                    $table = $table . '<td>'. $value .'</td>';
+                foreach($data as $key => $obs){
+
+                    $table = $table . '<tr>';
+                    
+                    foreach($obs as $key => $value){
+                        $table = $table . '<td>'. $value .'</td>';
+                    }
+                    
+                    $table = $table . 
+                    '<td>
+                        <a class="btn btn-primary" href="'.$path.'/'.$obs['id'].'?action=edit" role="button"><i class="far fa-edit"></i> Modifier</a>
+                        <a class="btn btn-danger" href="'.$path.'/'.$obs['id'].'?action=delete" role="button"><i class="far fa-trash-alt"></i> Supprimer</a>
+                    </td>';
+                    
+                    $table = $table . '</tr>';
                 }
                 
                 $table = $table . 
-                '<td>
-                    <a class="btn btn-primary" href="'.$path.'/'.$obs['id'].'?action=edit" role="button"><i class="far fa-edit"></i> Modifier</a>
-                    <a class="btn btn-danger" href="'.$path.'/'.$obs['id'].'?action=delete" role="button"><i class="far fa-trash-alt"></i> Supprimer</a>
-                </td>';
-                
-                $table = $table . '</tr>';
+                        '</tbody>
+                    </table>
+                </div>';
             }
-            
-            $table = $table . 
-                    '</tbody>
-                </table>
-            </div>';
-            
             return $table;
         }
 
-        public function generateForm($inputs, $actions = array('form'=>'/web', 'back' => '/web'), $data=null){        
-            
+        public function generateForm($inputs, $actions = array('form'=>'/web', 'back' => '/web'), $data=null, $hiddenInput = null){        
+           
             $form = null;
             $form = $form . '<form method="POST" action="'. $actions['form'] .'">';
             
+            if($hiddenInput!==null){
+                $form = $form . '<input type="hidden" name="'.$hiddenInput.'" value="'.$data[0][$hiddenInput].'">';
+            }
+
             foreach($inputs as $key => $value){
                 $attributes = $this->attributes[$key];
 
@@ -137,6 +149,8 @@ class ViewEngine {
 
             $form = $form . '</form>';
             return $form;
+
+            
         }
 
         public function generateNavbar($user = null, $active = null){
@@ -155,8 +169,8 @@ class ViewEngine {
             $title = 
             '<div class="btn-toolbar justify-content-between" role="toolbar" 
                 aria-label="Toolbar with button groups"
-                style="padding-top:60px;">
-                <h4>'. $str. '</h4>
+                style="padding-top:36px;">
+                <h2 style="padding:0;margin:0">'. $str. '</h2>
                 '.$btn.'
             </div><hr/>';
